refactor: simplify codebase following suckless principles
- Remove CallResponse::Unused and CastResponse::Unused variants
  Default implementations now panic with clear message
- Remove ChildPidWrapper, add impl HasPid for Pid instead
- Delete messages.rs module, use () instead of Unused type
- Reduce lib.rs exports, move supervisor internals to supervisor:: namespace
- Move gen_server tests to gen_server_tests.rs (1431 lines)
- Move supervisor tests to supervisor_tests.rs (768 lines)

Implementation file reductions:
- gen_server.rs: 2245 → 808 lines (-64%)
- supervisor.rs: 2111 → 1389 lines (-34%)
