--- concurrency/src/messages.rs
@@ -1,2 +1,0 @@
-#[derive(Clone, Debug)]
-pub struct Unused;
