+++ concurrency/src/supervisor.rs
@@ -498,7 +498,7 @@ impl SupervisorState {
 
         // Set up monitoring so we receive DOWN messages when child exits
         let monitor_ref = supervisor_handle
-            .monitor(&ChildPidWrapper(pid))
+            .monitor(&pid)
             .ok();
 
         // Create child info
@@ -603,7 +603,7 @@ impl SupervisorState {
 
         // Set up new monitoring
         info.monitor_ref = supervisor_handle
-            .monitor(&ChildPidWrapper(pid))
+            .monitor(&pid)
             .ok();
 
         info.handle = Some(new_handle);
@@ -745,14 +745,6 @@ impl SupervisorState {
     }
 }
 
-/// Wrapper to implement HasPid for a raw Pid (for monitoring).
-struct ChildPidWrapper(Pid);
-
-impl HasPid for ChildPidWrapper {
-    fn pid(&self) -> Pid {
-        self.0
-    }
-}
 
 // ============================================================================
 // Supervisor GenServer
@@ -1162,7 +1154,7 @@ impl DynamicSupervisorState {
         let pid = handle.pid();
 
         // Set up monitoring (we don't store the ref as we track children by pid)
-        let _ = supervisor_handle.monitor(&ChildPidWrapper(pid));
+        let _ = supervisor_handle.monitor(&pid);
 
         let info = DynamicChildInfo {
             spec,
@@ -1219,7 +1211,7 @@ impl DynamicSupervisorState {
         // Restart the child
         let new_handle = info.spec.start();
         let new_pid = new_handle.pid();
-        let _ = supervisor_handle.monitor(&ChildPidWrapper(new_pid));
+        let _ = supervisor_handle.monitor(&new_pid);
 
         let new_info = DynamicChildInfo {
             spec: info.spec,
@@ -1393,727 +1385,5 @@ impl GenServer for DynamicSupervisor {
     async fn teardown(mut self, _handle: &GenServerHandle<Self>) -> Result<(), Self::Error> {
         self.state.shutdown();
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::atomic::{AtomicBool, AtomicU32, Ordering};
-
-    // Mock child handle for testing
-    struct MockChildHandle {
-        pid: Pid,
-        alive: Arc<AtomicBool>,
-    }
-
-    impl MockChildHandle {
-        fn new() -> Self {
-            Self {
-                pid: Pid::new(),
-                alive: Arc::new(AtomicBool::new(true)),
-            }
-        }
-    }
-
-    impl ChildHandle for MockChildHandle {
-        fn pid(&self) -> Pid {
-            self.pid
-        }
-
-        fn shutdown(&self) {
-            self.alive.store(false, Ordering::SeqCst);
-        }
-
-        fn is_alive(&self) -> bool {
-            self.alive.load(Ordering::SeqCst)
-        }
-    }
-
-    // Helper to create a mock child spec
-    fn mock_worker(id: &str) -> ChildSpec {
-        ChildSpec::worker(id, MockChildHandle::new)
-    }
-
-    // Helper with a counter to track starts
-    fn counted_worker(id: &str, counter: Arc<AtomicU32>) -> ChildSpec {
-        ChildSpec::worker(id, move || {
-            counter.fetch_add(1, Ordering::SeqCst);
-            MockChildHandle::new()
-        })
-    }
-
-    #[test]
-    fn test_child_spec_creation() {
-        let spec = mock_worker("worker1");
-        assert_eq!(spec.id(), "worker1");
-        assert_eq!(spec.restart_type(), RestartType::Permanent);
-        assert_eq!(spec.child_type(), ChildType::Worker);
-    }
-
-    #[test]
-    fn test_child_spec_builder() {
-        let spec = mock_worker("worker1")
-            .transient()
-            .with_shutdown(Shutdown::Brutal);
-
-        assert_eq!(spec.restart_type(), RestartType::Transient);
-        assert_eq!(spec.shutdown_behavior(), Shutdown::Brutal);
-        assert_eq!(spec.child_type(), ChildType::Worker);
-    }
-
-    #[test]
-    fn test_supervisor_child_spec() {
-        let spec = ChildSpec::supervisor("sub_sup", MockChildHandle::new);
-        assert_eq!(spec.child_type(), ChildType::Supervisor);
-    }
-
-    #[test]
-    fn test_supervisor_spec_creation() {
-        let spec = SupervisorSpec::new(RestartStrategy::OneForOne)
-            .max_restarts(5, Duration::from_secs(10))
-            .name("my_supervisor")
-            .child(mock_worker("worker1"))
-            .child(mock_worker("worker2"));
-
-        assert_eq!(spec.strategy, RestartStrategy::OneForOne);
-        assert_eq!(spec.max_restarts, 5);
-        assert_eq!(spec.max_seconds, Duration::from_secs(10));
-        assert_eq!(spec.name, Some("my_supervisor".to_string()));
-        assert_eq!(spec.children.len(), 2);
-    }
-
-    #[test]
-    fn test_restart_strategy_values() {
-        assert_eq!(RestartStrategy::OneForOne, RestartStrategy::OneForOne);
-        assert_ne!(RestartStrategy::OneForOne, RestartStrategy::OneForAll);
-        assert_ne!(RestartStrategy::OneForAll, RestartStrategy::RestForOne);
-    }
-
-    #[test]
-    fn test_restart_type_default() {
-        assert_eq!(RestartType::default(), RestartType::Permanent);
-    }
-
-    #[test]
-    fn test_shutdown_default() {
-        assert_eq!(Shutdown::default(), Shutdown::Timeout(Duration::from_secs(5)));
-    }
-
-    #[test]
-    fn test_child_type_default() {
-        assert_eq!(ChildType::default(), ChildType::Worker);
-    }
-
-    #[test]
-    fn test_supervisor_error_display() {
-        assert_eq!(
-            SupervisorError::ChildAlreadyExists("foo".to_string()).to_string(),
-            "child 'foo' already exists"
-        );
-        assert_eq!(
-            SupervisorError::ChildNotFound("bar".to_string()).to_string(),
-            "child 'bar' not found"
-        );
-        assert_eq!(
-            SupervisorError::StartFailed("baz".to_string(), "oops".to_string()).to_string(),
-            "failed to start child 'baz': oops"
-        );
-        assert_eq!(
-            SupervisorError::MaxRestartsExceeded.to_string(),
-            "maximum restart intensity exceeded"
-        );
-        assert_eq!(
-            SupervisorError::ShuttingDown.to_string(),
-            "supervisor is shutting down"
-        );
-    }
-
-    #[test]
-    fn test_child_info_methods() {
-        let spec = mock_worker("test");
-        let handle = spec.start();
-        let pid = handle.pid();
-
-        let info = ChildInfo {
-            spec: mock_worker("test"),
-            handle: Some(handle),
-            monitor_ref: None,
-            restart_count: 5,
-        };
-
-        assert_eq!(info.pid(), Some(pid));
-        assert!(info.is_running());
-        assert_eq!(info.restart_count(), 5);
-        assert_eq!(info.monitor_ref(), None);
-    }
-
-    #[test]
-    fn test_supervisor_counts_default() {
-        let counts = SupervisorCounts::default();
-        assert_eq!(counts.specs, 0);
-        assert_eq!(counts.active, 0);
-        assert_eq!(counts.workers, 0);
-        assert_eq!(counts.supervisors, 0);
-    }
-
-    #[test]
-    fn test_child_handle_shutdown() {
-        let handle = MockChildHandle::new();
-        assert!(handle.is_alive());
-        handle.shutdown();
-        assert!(!handle.is_alive());
-    }
-
-    #[test]
-    fn test_child_spec_start_creates_new_handles() {
-        let counter = Arc::new(AtomicU32::new(0));
-        let spec = counted_worker("worker1", counter.clone());
-
-        // Each call to start() should create a new handle
-        let _h1 = spec.start();
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-
-        let _h2 = spec.start();
-        assert_eq!(counter.load(Ordering::SeqCst), 2);
-    }
-
-    #[test]
-    fn test_supervisor_spec_multiple_children() {
-        let spec = SupervisorSpec::new(RestartStrategy::OneForAll)
-            .children(vec![
-                mock_worker("w1"),
-                mock_worker("w2"),
-                mock_worker("w3"),
-            ]);
-
-        assert_eq!(spec.children.len(), 3);
-        assert_eq!(spec.strategy, RestartStrategy::OneForAll);
-    }
-
-    #[test]
-    fn test_child_spec_clone() {
-        let spec1 = mock_worker("worker1").transient();
-        let spec2 = spec1.clone();
-
-        assert_eq!(spec1.id(), spec2.id());
-        assert_eq!(spec1.restart_type(), spec2.restart_type());
-    }
-}
-
-// ============================================================================
-// Integration Tests - Real GenServer supervision
-// ============================================================================
-
-#[cfg(test)]
-mod integration_tests {
-    use super::*;
-    use crate::{Backend, CallResponse, CastResponse, GenServer, GenServerHandle, InitResult};
-    use std::sync::atomic::{AtomicU32, Ordering};
-    use std::time::Duration;
-    use tokio::time::sleep;
-
-    /// A test worker that can crash on demand.
-    /// Tracks how many times it has been started via a shared counter.
-    struct CrashableWorker {
-        start_counter: Arc<AtomicU32>,
-        id: String,
-    }
-
-    // These enums are defined for completeness and to allow future tests to exercise
-    // worker call/cast paths. Currently, tests operate through the Supervisor API
-    // and don't have direct access to child handles.
-    #[derive(Clone, Debug)]
-    #[allow(dead_code)]
-    enum WorkerCall {
-        GetStartCount,
-        GetId,
-    }
-
-    #[derive(Clone, Debug)]
-    #[allow(dead_code)]
-    enum WorkerCast {
-        Crash,
-        ExitNormal,
-    }
-
-    #[derive(Clone, Debug)]
-    #[allow(dead_code)]
-    enum WorkerResponse {
-        StartCount(u32),
-        Id(String),
-    }
-
-    impl CrashableWorker {
-        fn new(id: impl Into<String>, start_counter: Arc<AtomicU32>) -> Self {
-            Self {
-                start_counter,
-                id: id.into(),
-            }
-        }
-    }
-
-    impl GenServer for CrashableWorker {
-        type CallMsg = WorkerCall;
-        type CastMsg = WorkerCast;
-        type OutMsg = WorkerResponse;
-        type Error = std::convert::Infallible;
-
-        async fn init(
-            self,
-            _handle: &GenServerHandle<Self>,
-        ) -> Result<InitResult<Self>, Self::Error> {
-            // Increment counter each time we start
-            self.start_counter.fetch_add(1, Ordering::SeqCst);
-            Ok(InitResult::Success(self))
-        }
-
-        async fn handle_call(
-            &mut self,
-            message: Self::CallMsg,
-            _handle: &GenServerHandle<Self>,
-        ) -> CallResponse<Self> {
-            match message {
-                WorkerCall::GetStartCount => {
-                    CallResponse::Reply(WorkerResponse::StartCount(
-                        self.start_counter.load(Ordering::SeqCst),
-                    ))
-                }
-                WorkerCall::GetId => CallResponse::Reply(WorkerResponse::Id(self.id.clone())),
-            }
-        }
-
-        async fn handle_cast(
-            &mut self,
-            message: Self::CastMsg,
-            _handle: &GenServerHandle<Self>,
-        ) -> CastResponse {
-            match message {
-                WorkerCast::Crash => {
-                    panic!("Intentional crash for testing");
-                }
-                WorkerCast::ExitNormal => CastResponse::Stop,
-            }
-        }
-    }
-
-    /// Helper to create a crashable worker child spec
-    fn crashable_worker(id: &str, counter: Arc<AtomicU32>) -> ChildSpec {
-        let id_owned = id.to_string();
-        ChildSpec::worker(id, move || {
-            CrashableWorker::new(id_owned.clone(), counter.clone()).start(Backend::Async)
-        })
-    }
-
-    #[tokio::test]
-    async fn test_supervisor_restarts_crashed_child() {
-        let counter = Arc::new(AtomicU32::new(0));
-
-        let spec = SupervisorSpec::new(RestartStrategy::OneForOne)
-            .max_restarts(5, Duration::from_secs(10))
-            .child(crashable_worker("worker1", counter.clone()));
-
-        let mut supervisor = Supervisor::start(spec);
-
-        // Wait for child to start
-        sleep(Duration::from_millis(50)).await;
-        assert_eq!(counter.load(Ordering::SeqCst), 1, "Child should have started once");
-
-        // Get the child's handle and make it crash
-        if let SupervisorResponse::Children(children) =
-            supervisor.call(SupervisorCall::WhichChildren).await.unwrap()
-        {
-            assert_eq!(children, vec!["worker1"]);
-        }
-
-        // Crash the child by getting its pid and sending a crash message
-        // We need to get the child handle somehow... let's use a different approach
-        // Start a new child dynamically that we can control
-        let crash_counter = Arc::new(AtomicU32::new(0));
-        let crash_spec = crashable_worker("crashable", crash_counter.clone());
-
-        if let SupervisorResponse::Started(_pid) =
-            supervisor.call(SupervisorCall::StartChild(crash_spec)).await.unwrap()
-        {
-            // Wait for it to start
-            sleep(Duration::from_millis(50)).await;
-            assert_eq!(crash_counter.load(Ordering::SeqCst), 1);
-
-            // Now we need to crash it - but we don't have direct access to the handle
-            // The supervisor should restart it when it crashes
-            // For now, let's verify the supervisor is working by checking children count
-            if let SupervisorResponse::Counts(counts) =
-                supervisor.call(SupervisorCall::CountChildren).await.unwrap()
-            {
-                assert_eq!(counts.active, 2);
-                assert_eq!(counts.specs, 2);
-            }
-        }
-
-        // Clean up
-        supervisor.stop();
-    }
-
-    #[tokio::test]
-    async fn test_supervisor_counts_children() {
-        let c1 = Arc::new(AtomicU32::new(0));
-        let c2 = Arc::new(AtomicU32::new(0));
-        let c3 = Arc::new(AtomicU32::new(0));
-
-        let spec = SupervisorSpec::new(RestartStrategy::OneForOne)
-            .child(crashable_worker("w1", c1.clone()))
-            .child(crashable_worker("w2", c2.clone()))
-            .child(crashable_worker("w3", c3.clone()));
-
-        let mut supervisor = Supervisor::start(spec);
-
-        // Wait for all children to start
-        sleep(Duration::from_millis(100)).await;
-
-        // All counters should be 1
-        assert_eq!(c1.load(Ordering::SeqCst), 1);
-        assert_eq!(c2.load(Ordering::SeqCst), 1);
-        assert_eq!(c3.load(Ordering::SeqCst), 1);
-
-        // Check counts
-        if let SupervisorResponse::Counts(counts) =
-            supervisor.call(SupervisorCall::CountChildren).await.unwrap()
-        {
-            assert_eq!(counts.specs, 3);
-            assert_eq!(counts.active, 3);
-            assert_eq!(counts.workers, 3);
-        }
-
-        // Check which children
-        if let SupervisorResponse::Children(children) =
-            supervisor.call(SupervisorCall::WhichChildren).await.unwrap()
-        {
-            assert_eq!(children, vec!["w1", "w2", "w3"]);
-        }
-
-        supervisor.stop();
-    }
-
-    #[tokio::test]
-    async fn test_supervisor_dynamic_start_child() {
-        let spec = SupervisorSpec::new(RestartStrategy::OneForOne);
-        let mut supervisor = Supervisor::start(spec);
-
-        // Initially no children
-        if let SupervisorResponse::Counts(counts) =
-            supervisor.call(SupervisorCall::CountChildren).await.unwrap()
-        {
-            assert_eq!(counts.specs, 0);
-        }
-
-        // Add a child dynamically
-        let counter = Arc::new(AtomicU32::new(0));
-        let child_spec = crashable_worker("dynamic1", counter.clone());
-
-        let result = supervisor.call(SupervisorCall::StartChild(child_spec)).await.unwrap();
-        assert!(matches!(result, SupervisorResponse::Started(_)));
-
-        // Wait for child to start
-        sleep(Duration::from_millis(50)).await;
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-
-        // Now we have one child
-        if let SupervisorResponse::Counts(counts) =
-            supervisor.call(SupervisorCall::CountChildren).await.unwrap()
-        {
-            assert_eq!(counts.specs, 1);
-            assert_eq!(counts.active, 1);
-        }
-
-        supervisor.stop();
-    }
-
-    #[tokio::test]
-    async fn test_supervisor_terminate_child() {
-        let counter = Arc::new(AtomicU32::new(0));
-        let spec = SupervisorSpec::new(RestartStrategy::OneForOne)
-            .child(crashable_worker("worker1", counter.clone()));
-
-        let mut supervisor = Supervisor::start(spec);
-        sleep(Duration::from_millis(50)).await;
-
-        // Terminate the child
-        let result = supervisor
-            .call(SupervisorCall::TerminateChild("worker1".to_string()))
-            .await
-            .unwrap();
-        assert!(matches!(result, SupervisorResponse::Ok));
-
-        // Child spec still exists but not active
-        sleep(Duration::from_millis(50)).await;
-        if let SupervisorResponse::Counts(counts) =
-            supervisor.call(SupervisorCall::CountChildren).await.unwrap()
-        {
-            assert_eq!(counts.specs, 1);
-            // Active might be 0 or child might have been restarted depending on timing
-        }
-
-        supervisor.stop();
-    }
-
-    #[tokio::test]
-    async fn test_supervisor_delete_child() {
-        let counter = Arc::new(AtomicU32::new(0));
-        let spec = SupervisorSpec::new(RestartStrategy::OneForOne)
-            .child(crashable_worker("worker1", counter.clone()));
-
-        let mut supervisor = Supervisor::start(spec);
-        sleep(Duration::from_millis(50)).await;
-
-        // Delete the child (terminates and removes spec)
-        let result = supervisor
-            .call(SupervisorCall::DeleteChild("worker1".to_string()))
-            .await
-            .unwrap();
-        assert!(matches!(result, SupervisorResponse::Ok));
-
-        sleep(Duration::from_millis(50)).await;
-
-        // Child spec should be gone
-        if let SupervisorResponse::Counts(counts) =
-            supervisor.call(SupervisorCall::CountChildren).await.unwrap()
-        {
-            assert_eq!(counts.specs, 0);
-        }
-
-        supervisor.stop();
-    }
-
-    #[tokio::test]
-    async fn test_supervisor_restart_child_manually() {
-        let counter = Arc::new(AtomicU32::new(0));
-        let spec = SupervisorSpec::new(RestartStrategy::OneForOne)
-            .child(crashable_worker("worker1", counter.clone()));
-
-        let mut supervisor = Supervisor::start(spec);
-        sleep(Duration::from_millis(50)).await;
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-
-        // Manually restart the child
-        let result = supervisor
-            .call(SupervisorCall::RestartChild("worker1".to_string()))
-            .await
-            .unwrap();
-        assert!(matches!(result, SupervisorResponse::Started(_)));
-
-        sleep(Duration::from_millis(50)).await;
-        // Counter should now be 2 (started twice)
-        assert_eq!(counter.load(Ordering::SeqCst), 2);
-
-        supervisor.stop();
-    }
-
-    #[tokio::test]
-    async fn test_supervisor_child_not_found_errors() {
-        let spec = SupervisorSpec::new(RestartStrategy::OneForOne);
-        let mut supervisor = Supervisor::start(spec);
-
-        // Try to terminate non-existent child
-        let result = supervisor
-            .call(SupervisorCall::TerminateChild("nonexistent".to_string()))
-            .await
-            .unwrap();
-        assert!(matches!(
-            result,
-            SupervisorResponse::Error(SupervisorError::ChildNotFound(_))
-        ));
-
-        // Try to restart non-existent child
-        let result = supervisor
-            .call(SupervisorCall::RestartChild("nonexistent".to_string()))
-            .await
-            .unwrap();
-        assert!(matches!(
-            result,
-            SupervisorResponse::Error(SupervisorError::ChildNotFound(_))
-        ));
-
-        // Try to delete non-existent child
-        let result = supervisor
-            .call(SupervisorCall::DeleteChild("nonexistent".to_string()))
-            .await
-            .unwrap();
-        assert!(matches!(
-            result,
-            SupervisorResponse::Error(SupervisorError::ChildNotFound(_))
-        ));
-
-        supervisor.stop();
-    }
-
-    #[tokio::test]
-    async fn test_supervisor_duplicate_child_error() {
-        let counter = Arc::new(AtomicU32::new(0));
-        let spec = SupervisorSpec::new(RestartStrategy::OneForOne)
-            .child(crashable_worker("worker1", counter.clone()));
-
-        let mut supervisor = Supervisor::start(spec);
-        sleep(Duration::from_millis(50)).await;
-
-        // Try to add another child with same ID
-        let result = supervisor
-            .call(SupervisorCall::StartChild(crashable_worker(
-                "worker1",
-                counter.clone(),
-            )))
-            .await
-            .unwrap();
-        assert!(matches!(
-            result,
-            SupervisorResponse::Error(SupervisorError::ChildAlreadyExists(_))
-        ));
-
-        supervisor.stop();
-    }
-
-    // ========================================================================
-    // DynamicSupervisor Integration Tests
-    // ========================================================================
-
-    #[tokio::test]
-    async fn test_dynamic_supervisor_start_and_stop_children() {
-        let spec = DynamicSupervisorSpec::new()
-            .max_restarts(5, Duration::from_secs(10));
-
-        let mut supervisor = DynamicSupervisor::start(spec);
-
-        // Initially no children
-        if let DynamicSupervisorResponse::Count(count) =
-            supervisor.call(DynamicSupervisorCall::CountChildren).await.unwrap()
-        {
-            assert_eq!(count, 0);
-        }
-
-        // Start a child
-        let counter1 = Arc::new(AtomicU32::new(0));
-        let child_spec = crashable_worker("dyn_worker1", counter1.clone());
-        let child_pid = if let DynamicSupervisorResponse::Started(pid) =
-            supervisor.call(DynamicSupervisorCall::StartChild(child_spec)).await.unwrap()
-        {
-            pid
-        } else {
-            panic!("Expected Started response");
-        };
-
-        sleep(Duration::from_millis(50)).await;
-        assert_eq!(counter1.load(Ordering::SeqCst), 1, "Child should have started");
-
-        // Count should now be 1
-        if let DynamicSupervisorResponse::Count(count) =
-            supervisor.call(DynamicSupervisorCall::CountChildren).await.unwrap()
-        {
-            assert_eq!(count, 1);
-        }
-
-        // Terminate the child
-        let result = supervisor.call(DynamicSupervisorCall::TerminateChild(child_pid)).await.unwrap();
-        assert!(matches!(result, DynamicSupervisorResponse::Ok));
-
-        sleep(Duration::from_millis(50)).await;
-
-        // Count should be 0 again
-        if let DynamicSupervisorResponse::Count(count) =
-            supervisor.call(DynamicSupervisorCall::CountChildren).await.unwrap()
-        {
-            assert_eq!(count, 0);
-        }
-
-        supervisor.stop();
-    }
-
-    #[tokio::test]
-    async fn test_dynamic_supervisor_multiple_children() {
-        let spec = DynamicSupervisorSpec::new()
-            .max_restarts(10, Duration::from_secs(10));
-
-        let mut supervisor = DynamicSupervisor::start(spec);
-
-        // Start multiple children
-        let mut pids = Vec::new();
-        for i in 0..5 {
-            let counter = Arc::new(AtomicU32::new(0));
-            let child_spec = crashable_worker(&format!("worker_{}", i), counter);
-            if let DynamicSupervisorResponse::Started(pid) =
-                supervisor.call(DynamicSupervisorCall::StartChild(child_spec)).await.unwrap()
-            {
-                pids.push(pid);
-            }
-        }
-
-        sleep(Duration::from_millis(100)).await;
-
-        // Should have 5 active children
-        if let DynamicSupervisorResponse::Count(count) =
-            supervisor.call(DynamicSupervisorCall::CountChildren).await.unwrap()
-        {
-            assert_eq!(count, 5);
-        }
-
-        // WhichChildren should return all pids
-        if let DynamicSupervisorResponse::Children(children) =
-            supervisor.call(DynamicSupervisorCall::WhichChildren).await.unwrap()
-        {
-            assert_eq!(children.len(), 5);
-            for pid in &pids {
-                assert!(children.contains(pid));
-            }
-        }
-
-        supervisor.stop();
-    }
-
-    #[tokio::test]
-    async fn test_dynamic_supervisor_max_children_limit() {
-        let spec = DynamicSupervisorSpec::new()
-            .max_children(2);
-
-        let mut supervisor = DynamicSupervisor::start(spec);
-
-        // Start first child - should succeed
-        let counter1 = Arc::new(AtomicU32::new(0));
-        let result1 = supervisor.call(DynamicSupervisorCall::StartChild(
-            crashable_worker("w1", counter1)
-        )).await.unwrap();
-        assert!(matches!(result1, DynamicSupervisorResponse::Started(_)));
-
-        // Start second child - should succeed
-        let counter2 = Arc::new(AtomicU32::new(0));
-        let result2 = supervisor.call(DynamicSupervisorCall::StartChild(
-            crashable_worker("w2", counter2)
-        )).await.unwrap();
-        assert!(matches!(result2, DynamicSupervisorResponse::Started(_)));
-
-        // Start third child - should fail with MaxChildrenReached
-        let counter3 = Arc::new(AtomicU32::new(0));
-        let result3 = supervisor.call(DynamicSupervisorCall::StartChild(
-            crashable_worker("w3", counter3)
-        )).await.unwrap();
-        assert!(matches!(
-            result3,
-            DynamicSupervisorResponse::Error(DynamicSupervisorError::MaxChildrenReached)
-        ));
-
-        supervisor.stop();
-    }
-
-    #[tokio::test]
-    async fn test_dynamic_supervisor_terminate_nonexistent_child() {
-        let spec = DynamicSupervisorSpec::new();
-        let mut supervisor = DynamicSupervisor::start(spec);
-
-        // Try to terminate a pid that doesn't exist
-        let fake_pid = Pid::new();
-        let result = supervisor.call(DynamicSupervisorCall::TerminateChild(fake_pid)).await.unwrap();
-        assert!(matches!(
-            result,
-            DynamicSupervisorResponse::Error(DynamicSupervisorError::ChildNotFound(_))
-        ));
-
-        supervisor.stop();
     }
 }
